chore(discord): decompose intake_turn.rs under giant baseline (#3837)
Behavior-preserving decomposition of the ~3.5k-line `handle_text_message`
monolith. Three cohesive clusters are lifted VERBATIM out of the function
into sibling `intake_turn/` submodules; the only edits are the function
wrapper, parameter threading, and path/visibility plumbing — no logic,
signature (of existing fns), or runtime-behavior change.

Extraction map:
- `voice_intake::resolve_intake_voice_announcement` — the
  `let voice_announcement = ...` resolution expression + its
  `voice_announcement_already_accepted` flag (returns the tuple).
- `race_loss::handle_race_loss_enqueue` — the entire `if !started { .. }`
  race-loss path (mailbox enqueue + queued-placeholder render + queue-pending
  reaction lifecycle); its `return Ok(())`/`return Err(..)` map 1:1, so the
  caller does `if !started { return handle_race_loss_enqueue(..).await; }`.
- `turn_watchdog::spawn_text_turn_watchdog` — the per-turn watchdog spawn
  block (deadline/ceiling compute + `text_turn_watchdog` observed task).

intake_turn.rs production LoC: 3616 -> 2797 (well under the 3722 giant
baseline, which is lowered to 2797 to lock in the win). New modules are
544/264/155 LoC, all under the new `intake_turn/**` 700-line namespace cap.

Gates: cargo fmt --check clean; cargo check --lib clean (no new warnings);
`cargo test --lib message_handler` 66/66 green; agent-maintenance freshness
check passed; audit_maintainability --check exit 0; hotfile ratchet exit 0
(intake_turn.rs is NOT a #3016 hotfile).
