+++ ARCHITECTURE.md
@@ -523,6 +523,10 @@ src/
 │   │   │   │   ├── reaction_remove.rs
 │   │   │   │   └── stale_turn.rs
 │   │   │   ├── message_handler/
+│   │   │   │   ├── intake_turn/
+│   │   │   │   │   ├── race_loss.rs
+│   │   │   │   │   ├── turn_watchdog.rs
+│   │   │   │   │   └── voice_intake.rs
 │   │   │   │   ├── attachments.rs
 │   │   │   │   ├── control.rs
 │   │   │   │   ├── goal_lifecycle.rs
